tetris move + general refactor
- moved tetris to its own namespace
- deteted bunch of global states and `externs`
- readded score return from apps
