--- src/colors.hpp
@@ -1,7 +1,0 @@
-#ifndef COLORS_HPP
-#define COLORS_HPP
-#include <array>
-#include <Logic.hpp>
-extern std::array< Rgb, 8 > colors;
-extern std::array< Rgb, 10 > number_colors;
-#endif
