--- src/global_vars.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-#include <stdlib.h>
-#include <Logic.hpp>
-
-extern int brightness;
-extern int view_offset;
