+++ src/input.hpp
@@ -3,6 +3,13 @@
 #include <cstdint>
 #include "utils.hpp"
 
+enum class dpad_direction : uint8_t {
+    up = 0,
+    right,
+    down,
+    left,
+};
+
 struct joystick_event {
     Vec3 current_input; // processed input, should be used for game input
     Vec3 raw_input; // raw input received from joystick
@@ -10,8 +17,8 @@ struct joystick_event {
     // emulated dpad, bitfield for if dirs are out of their deadzone (bit 0 to 3, top, right, bottom, left)
     uint8_t dpad_state;
 
-    inline bool is_dpad_active(int dir_index) {
-        return (dpad_state >> dir_index) & 1;
+    inline bool is_dpad_active(dpad_direction dir) const {
+        return (dpad_state >> static_cast<uint8_t>(dir)) & 1;
     }
 };
