+++ src/shanke/shanke.hpp
@@ -3,5 +3,5 @@
 
 namespace shanke {
     void shanke_input_cb(const joystick_event& ev);
-    void shanke_main();
+    int shanke_main();
 }
